+++ CHANGELOG.rst
@@ -15,6 +15,8 @@ v0.8 (XXXX/XX/XX)
 -----------------
 Simon Gravelle, Philip Loche, Marc Sauter, Henrik Stooß, Philipp Staerk, Adyant Agrawal
 
+- Rename radial distribution function to pair distribution function (!278)
+- Add RDF derivation and explain role of dz. (!278)
 - Implement 1D pair distribution function in RDFCylinder (!276)
 - Sort format and add more atomtypes to ``atomtypes.dat`` (!291)
 - Add grouping option to `DipoleAngle` module (!290)